Streams should check for Session shutdown when waiting for data
Stream.Close is called when the Session's keepalive fails. This call wakes up
Stream.Read & Stream.write waiters. The wakeup is via a non-blocking send to
recvNotifyCh/sendNotifyCh, which are unbuffered channels.

In Stream.Read this send can be attempted between the call to
s.stateLock.Unlock() and the select statement at the bottom of the
method. When this occurs, if deadlines haven't been set on the Stream in
question, the call to Stream.Read will block forever.

By adding Session.shutdownCh to the channels in the select statement,
this coordination gap is addressed.

This uses `goto START` instead of `return 0, io.EOF`, as that would potentially
leave a last read on the buffer in the Read case.